Changed 'simulation' to 'core'. Merged similar modules in core.
