+++ nomad/adapt/optimal_spawn.py
@@ -15,10 +15,10 @@ child(s',ti) <------------- child(s',ts)
 2. If parent_i is coupled to another.
 """
 import numpy as np
-import nomad.simulation.glbl as glbl
-import nomad.simulation.log as log
-import nomad.simulation.step as step
-import nomad.simulation.evaluate as evaluate
+import nomad.core.glbl as glbl
+import nomad.core.log as log
+import nomad.core.step as step
+import nomad.core.surface as evaluate
 import nomad.adapt.utilities as utils
 
 coup_hist = []
