+++ nomad/interfaces/vibronic.py
@@ -4,9 +4,9 @@ Routines for running a vibronic coupling calculation.
 import numpy as np
 import scipy.linalg as sp_linalg
 import nomad.math.constants as constants
-import nomad.simulation.glbl as glbl
-import nomad.simulation.log as log
-import nomad.simulation.surface as surface
+import nomad.core.glbl as glbl
+import nomad.core.log as log
+import nomad.core.surface as surface
 
 
 ham = None
