+++ nomad/propagators/bulirsch_stoer.py
@@ -19,9 +19,9 @@ values of n.
 For more details, see http://apps.nrbook.com/fortran/index.html.
 """
 import numpy as np
-import nomad.simulation.glbl as glbl
-import nomad.simulation.timings as timings
-import nomad.simulation.evaluate as evaluate
+import nomad.core.glbl as glbl
+import nomad.core.timings as timings
+import nomad.core.surface as evaluate
 
 propphase = glbl.propagate['phase_prop']
