+++ nomad/propagators/rkf45.py
@@ -30,9 +30,9 @@ step size sfac*dt. Otherwise, use sfac*dt for next step. Update
 position, momentum using x4, p4.
 """
 import numpy as np
-import nomad.simulation.glbl as glbl
-import nomad.simulation.timings as timings
-import nomad.simulation.evaluate as evaluate
+import nomad.core.glbl as glbl
+import nomad.core.timings as timings
+import nomad.core.surface as evaluate
 
 
 rk_ordr = 6
